pull out Deck and FlashCard into /models
