+++ lib/main.dart
@@ -1,117 +1,10 @@
 import 'package:flutter/material.dart';
-import 'dart:convert';
-
 import 'package:flutter/rendering.dart';
 
+import './models/FlashCard.dart';
+import './models/Deck.dart';
+
 void main() => runApp(MyApp());
-
-enum FlashCardSide { front, back }
-
-class Deck {
-  Deck(this.id, this.title, this.cards, this.tags);
-
-  String id;
-  String title;
-  List<FlashCard> cards;
-  List<String> tags;
-
-  String toString() {
-    return "$id ${cards.toString()}";
-  }
-
-  Deck.fromJsonString(String jsonString) {
-    final Map<String, dynamic> jsonData = json.decode(jsonString);
-    id = jsonData["id"];
-    title = jsonData["title"];
-    cards = List.from(
-        jsonData["cards"].map((card) => new FlashCard.fromObject(card)));
-    tags = List.from(jsonData["tags"]);
-  }
-
-  Map<String, dynamic> toJson() {
-    return {
-      "id": id,
-      "title": title,
-      "tags": tags.toString(),
-      "cards": cards.toString()
-    };
-  }
-
-  List<FlashCard> getCards() {
-    return cards;
-  }
-}
-
-class FlashCard {
-  FlashCard(this.id, this.title, this.front, this.back, this.tags, this.decks);
-
-  String id;
-  String title;
-  String front;
-  String back;
-  List<String> tags;
-  List<String> decks;
-
-  String toString() {
-    return "$id $title $front $back ${tags.toString()}, ${decks.toString()}";
-  }
-
-  FlashCard.fromJsonString(String jsonString) {
-    final Map<String, dynamic> jsonData = json.decode(jsonString);
-    id = jsonData["id"];
-    title = jsonData["title"];
-    front = jsonData["front"];
-    back = jsonData["back"];
-    tags = List.from(jsonData["tags"]);
-    decks = List.from(jsonData["decks"]);
-  }
-
-  FlashCard.fromObject(flashcardObj) {
-    id = flashcardObj["id"];
-    title = flashcardObj["title"];
-    front = flashcardObj["front"];
-    back = flashcardObj["back"];
-    tags = List.from(flashcardObj["tags"]);
-    decks = List.from(flashcardObj["decks"]);
-  }
-
-  Map<String, dynamic> toJson() {
-    return {
-      "id": id,
-      "title": title,
-      "front": front,
-      "back": back,
-      "tags": tags.toString(),
-      "decks": decks.toString()
-    };
-  }
-
-  FractionallySizedBox toWidget({FlashCardSide sideToDisplay}) =>
-      FractionallySizedBox(
-        widthFactor: 0.8,
-        heightFactor: 0.8,
-        child: Card(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              new Expanded(
-                  flex: 1,
-                  child: Center(
-                    child: SingleChildScrollView(
-                        padding: EdgeInsets.all(24.0),
-                        child: Text(
-                          sideToDisplay == FlashCardSide.front ? front : back,
-                          style: TextStyle(
-                              fontWeight: FontWeight.w500, fontSize: 32),
-                          textAlign: TextAlign.center,
-                        )),
-                  ))
-            ],
-          ),
-        ),
-      );
-}
 
 class MyApp extends StatelessWidget {
   // This widget is the root of your application.
